requirements.txt
fastapi
requests
pydantic
python-dotenv
python-multipart
uvicorn
psycopg2-binary
rdflib
thefuzz
SPARQLWrapper
yt_dlp
spacy
beautifulsoup4
asyncio
aiohttp
asyncpg

